requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
python-dotenv==1.0.0
supabase==2.3.4
Pillow==10.2.0
pytz==2024.1
apscheduler==3.10.4
openai==1.12.0
